Expose methods for loading history from files as WorkflowExecutionHistoryLoader
