+++ temporal-testing/build.gradle
@@ -23,7 +23,7 @@ dependencies {
 
     junit4Api 'junit:junit:4.13.2'
 
-    junit5Api platform('org.junit:junit-bom:5.8.1')
+    junit5Api platform('org.junit:junit-bom:5.8.2')
     junit5Api 'org.junit.jupiter:junit-jupiter-api'
 
     testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter'
